VEA-12 Reorganized tests.
Moved the auth-related tests to their own test package, and kept one
customer-related test for the customer package.
